Zoo Decoration Functionality
I added the ability to decorate your zoo. currentDecorations is now a variable in localStorage, as is decorChoice. Currently there are no new assets so the things you decorate it with look the same, but I added a magenta border to the item when in decoration mode for debugging. Currently all you can do is place decorations in the pre-defined slots, I don't have time to do the click and dragging I was hoping for but this is fine for now.